+++ VerwijsIndexInterface/IProvider.cs
@@ -298,7 +298,7 @@ namespace Denion.WebService.VerwijsIndex
         /// Betalingskenmerk van de Provider,verplicht
         /// </summary>
         [DataMember(Order = 1, IsRequired = true)]
-        public long? PaymentAuthorisationId { get; set; }
+        public int? PaymentAuthorisationId { get; set; }
 
         /// <summary>
         /// Voertuigkenteken, verplicht
@@ -386,7 +386,7 @@ namespace Denion.WebService.VerwijsIndex
         /// Betalingskenmerk van de Provider,verplicht
         /// </summary>
         [DataMember(Order = 1, IsRequired = true)]
-        public long? PaymentAuthorisationId { get; set; }
+        public int? PaymentAuthorisationId { get; set; }
 
         /// <summary>
         /// Wordt er toegang verleend (of juist niet), optioneel
@@ -455,7 +455,7 @@ namespace Denion.WebService.VerwijsIndex
         /// Betalingskenmerk van de Provider,verplicht
         /// </summary>
         [DataMember(Order = 1, IsRequired = true)]
-        public long? PaymentAuthorisationId { get; set; }
+        public int? PaymentAuthorisationId { get; set; }
 
         /// <summary>
         /// Voertuigkenteken, verplicht
